Extracts duplicate test helpers into corresponding scopes.
